+++ glaze-lib/main.rkt
@@ -2,6 +2,8 @@
 
 (require "server.rkt"
          "api.rkt"
+         "api-macros.rkt"
+         "events.rkt"
          "browser.rkt"
          "assets.rkt"
          "build.rkt"
@@ -9,7 +11,8 @@
          "tray/main.rkt"
          "webview/main.rkt")
 
-(provide (all-from-out "server.rkt" "api.rkt" "browser.rkt" "assets.rkt" "build.rkt")
+(provide (all-from-out "server.rkt" "api.rkt" "api-macros.rkt" "events.rkt"
+                       "browser.rkt" "assets.rkt" "build.rkt")
          (all-from-out "app.rkt")
          (all-from-out "tray/main.rkt")
          (all-from-out "webview/main.rkt"))